backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
yfinance==0.2.36
python-multipart==0.0.9